+++ CHANGELOG.md
@@ -21,6 +21,7 @@ apiVersion2: v1
 -                    port: 5432
 ```
 - Yaml files path service: now you can use curly braces to find file, e.g.: `services.{yml,yaml}`
+- Yaml alphabetical: fixed right sort yaml by depth
 
 ## [5.1.2] - 2020-02-02
 ### Fixed